Add files via upload
we have created a web page for zero hunger. The basic use of this is to provide food who are in need. And we have created a donor page so that the donors can keep their data and send the detailed information about food. So that all the information given by the donors are inserted to data base . here we also created a receivers page, in that receiver fill his details and his organization address. by filling the receiver address the donor address which is near to the receiver address shows to the receiver so that receiver can receive food. and we have create feedback page so that the donors can no that the food is good or not ..,if the receivers have struggled by their food